requirements.txt
requests>=2.31
websockets>=12.0
python-dotenv>=1.0
urllib3>=1.26
cryptography>=42.0
